Move AccentShape to component folder
